Backend/requirements.txt
google-generativeai>=0.8.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-jose>=3.3.0
passlib>=1.7.4
